v0.30.1

+++ WpfApplication1/AdminWindows/UsersControlWindow.xaml.cs
@@ -17,40 +17,36 @@ namespace WpfApplication1
     public partial class UsersControlWindow : Window
     {
         string idText;
-        List<User> user = new List<User> { };
         public UsersControlWindow(string id)
         {
             InitializeComponent();
-            user = DataBase.GetUser();
-            dataGridUserOut.ItemsSource = user;
+            dataGridUserOut.ItemsSource = DataBase.GetUser();
             DataBase.SetLog(id, 0, 0, "Заполнение таблицы пользователей...");
             idText = id;
         }
 
         private void UserAdd_Click(object sender, RoutedEventArgs e)
         {
-            new UsersControlAddWindow(idText).ShowDialog();            
+            UsersControlAddWindow window = new UsersControlAddWindow(idText);
+            window.ShowDialog();
+            if (window.flag)
+            {
+                dataGridUserOut.ItemsSource = DataBase.GetUser();
+            }
         }
 
         private void UserDelete_Click(object sender, RoutedEventArgs e)
         {
-             try
-            {
-                if (dataGridUserOut.SelectedIndex != -1)
+
+            if (dataGridUserOut.SelectedIndex != -1 && Converter.DGCellToStringConvert(dataGridUserOut.SelectedIndex, 0, dataGridUserOut) != "Администратор")
                 {
                     string login = Converter.DGCellToStringConvert(dataGridUserOut.SelectedIndex, 1, dataGridUserOut);
                     DataBase.Query(new string[] { "@_login", "@_password" }
                         , new string[2] { login, Converter.DGCellToStringConvert(dataGridUserOut.SelectedIndex, 2, dataGridUserOut) }
                         , "DELETE FROM `user` WHERE U_NAME=@_login AND U_PASS=@_password;");
-                    user.RemoveAt(dataGridUserOut.SelectedIndex);
-                    dataGridUserOut.Items.Refresh();
+                    dataGridUserOut.ItemsSource = DataBase.GetUser();
                     DataBase.SetLog(idText, 1, 3, "Удаление пользователя,параметры:|логин:" + login + "|");
                 }
-            }
-            catch (Exception ex)
-            {
-                MessageBox.Show(ex.Message);
-            }
         }
 
         private void buttonUserEdit_Click(object sender, RoutedEventArgs e)
@@ -59,7 +55,12 @@ namespace WpfApplication1
             {
                 if (dataGridUserOut.SelectedIndex != -1)
                 {
-                    new UsersControlEditWindow(idText,Converter.DGCellToStringConvert(dataGridUserOut.SelectedIndex, 0, dataGridUserOut)).ShowDialog();
+                    UsersControlEditWindow window = new UsersControlEditWindow(idText,Converter.DGCellToStringConvert(dataGridUserOut.SelectedIndex, 0, dataGridUserOut));
+                    window.ShowDialog();
+                    if(window.flag)
+                    {
+                        dataGridUserOut.ItemsSource = DataBase.GetUser();
+                    }
                 }
             }
             catch (Exception ex)
@@ -78,6 +79,5 @@ namespace WpfApplication1
             dataGridUserOut.ItemsSource = DataBase.GetUser();
             DataBase.SetLog(idText, 1, 0, "Обновление таблицы пользователей...");
         }
-
     }
 }
